refactor(sidebar): extract connection pipeline into operations/pipeline.rs
